Move old booking utils
