+++ server/utils/outOfServiceBedUtils.test.ts
@@ -5,6 +5,7 @@ import {
   allOutOfServiceBedsTableHeaders,
   allOutOfServiceBedsTableRows,
   bedRevisionDetails,
+  generateConflictBespokeError,
   outOfServiceBedActions,
   outOfServiceBedTableHeaders,
   outOfServiceBedTableRows,
@@ -15,6 +16,7 @@ import {
 import { Cas1OutOfServiceBedSortField as OutOfServiceBedSortField } from '../@types/shared'
 import { sortHeader } from './sortHeader'
 import paths from '../paths/manage'
+import { SanitisedError } from '../sanitisedError'
 
 describe('outOfServiceBedUtils', () => {
   describe('allOutOfServiceBedsTableHeaders', () => {
@@ -238,6 +240,75 @@ describe('outOfServiceBedUtils', () => {
       expect(overwriteOoSBedWithUserInput(userInput, outOfServiceBed)).toEqual(
         expect.objectContaining({ referenceNumber: 'new reason' }),
       )
+    })
+  })
+
+  describe('generateConflictBespokeError', () => {
+    const premisesId = 'premises-id'
+    const bookingId = 'booking-id'
+    const bedId = 'bed-id'
+    const lostBedId = 'lost-bed-id'
+
+    it('generates a bespoke error when there is a conflicting booking', () => {
+      const err = {
+        data: {
+          detail: `Conflicting Booking: ${bookingId}`,
+        },
+      }
+
+      expect(generateConflictBespokeError(err as SanitisedError, premisesId, 'plural', bedId)).toEqual({
+        errorTitle: 'This bedspace is not available for the dates entered',
+        errorSummary: [
+          {
+            html: `They conflict with an <a href="${paths.premises.placements.show({
+              premisesId,
+              placementId: bookingId,
+            })}">existing booking</a>`,
+          },
+        ],
+      })
+    })
+
+    it('generates a bespoke error when there is a conflicting out-of-service bed', () => {
+      const err = {
+        data: {
+          detail: `An out-of-service bed already exists for dates from 2024-10-05 to 2024-10-20 which overlaps with the desired dates: ${lostBedId}`,
+        },
+      }
+
+      expect(generateConflictBespokeError(err as SanitisedError, premisesId, 'plural', bedId)).toEqual({
+        errorTitle: 'Out of service bed record cannot be created for the dates entered',
+        errorSummary: [
+          {
+            html: `They conflict with an <a href="${paths.outOfServiceBeds.show({
+              premisesId,
+              bedId,
+              id: lostBedId,
+              tab: 'details',
+            })}">existing out of service beds record</a>`,
+          },
+        ],
+      })
+    })
+
+    it('generates a bespoke error for a single date', () => {
+      const err = {
+        data: {
+          detail: `Conflicting Booking: ${bookingId}`,
+        },
+      }
+
+      expect(generateConflictBespokeError(err as SanitisedError, premisesId, 'singular', bedId)).toEqual({
+        errorTitle: 'This bedspace is not available for the date entered',
+        errorSummary: [
+          {
+            html: `It conflicts with an <a href="${paths.premises.placements.show({
+              premisesId,
+              placementId: bookingId,
+            })}">existing booking</a>`,
+          },
+        ],
+      })
     })
   })
 })
